Add feed API wrapper and move mergefeed into it
